Create kernels.algorithms package for QuantumKernelTrainer. Fix all CI/CD errors
