+++ .pylintdict
@@ -179,6 +179,7 @@ utils
 variational
 vec
 vqc
+vqe
 vqr
 vx
 vy
