+++ qiskit_machine_learning/kernels/quantum_kernel.py
@@ -136,7 +136,7 @@ class QuantumKernel:
         A property used by the ``RuntimeEncoder`` and ``RuntimeDecoder`` to serialize this class.
 
         Users who wish to serialize a ``QuantumKernel`` must ensure the feature map is
-        serializable via ``RuntimeEncoder``.
+        supported by Qiskit ``RuntimeEncoder``.
         """
         # Ensure user parameters are in a list, not a ParameterVector
         user_parameters_list = list(self._user_parameters)
@@ -153,14 +153,17 @@ class QuantumKernel:
     ) -> None:
         """
         Assign user parameters in the ``QuantumKernel`` feature map.
+
         Args:
             values (dict or iterable): Either a dictionary or iterable specifying the new
-                parameter values. If a dict, it specifies the mapping from ``current_parameter`` to
-                ``new_parameter``, where ``new_parameter`` can be a parameter expression or a
-                numeric value. If an iterable, the elements are assigned to the existing parameters
-                in the order of ``QuantumKernel.user_parameters``.
+            parameter values. If a dict, it specifies the mapping from ``current_parameter`` to
+            ``new_parameter``, where ``new_parameter`` can be a parameter expression or a
+            numeric value. If an iterable, the elements are assigned to the existing parameters
+            in the order of ``QuantumKernel.user_parameters``.
+
         Raises:
             ValueError: Incompatible number of user parameters and values
+
         """
         if self._user_parameters is None:
             raise ValueError(
