+++ qiskit_machine_learning/utils/loss_functions/__init__.py
@@ -13,7 +13,7 @@
 """
 Loss Functions (:mod:`qiskit_machine_learning.utils.loss_functions`)
 ====================================================================
-A collections of common loss functions to be used with the classifiers and regressors provided
+A collection of common loss functions to be used with the classifiers and regressors provided
 by Qiskit Machine Learning.
 
 .. currentmodule:: qiskit_machine_learning.utils.loss_functions
@@ -40,18 +40,17 @@ Loss Functions
    CrossEntropyLoss
    CrossEntropySigmoidLoss
    SVCLoss
-
 """
 
 from .loss_functions import (
     Loss,
-    KernelLoss,
     L1Loss,
     L2Loss,
     CrossEntropyLoss,
     CrossEntropySigmoidLoss,
-    SVCLoss,
 )
+
+from .kernel_loss_functions import KernelLoss, SVCLoss
 
 __all__ = [
     "Loss",
